# swagger.yaml
openapi: 3.0.0
info:
  title: SMS
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /subjects/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Subjects
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Subjects
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subjects"
    post:
      summary: CRUD ACTION update
      tags:
        - Subjects
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subjects"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Subjects"
  /subjects/findBy_Class/{key}:
    get:
      summary: CRUD ACTION findBy_Class
      tags:
        - Subjects
      parameters:
        - name: class
          in: path
          description: Class
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subjects"
  /subjects/findBy_Teachers/{key}:
    get:
      summary: CRUD ACTION findBy_Teachers
      tags:
        - Subjects
      parameters:
        - name: firstname
          in: path
          description: First Name
          required: true
          schema:
            type: string
        - name: lastname
          in: path
          description: Last Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subjects"
...
  /subjects/{id}/get_Class:
    get:
      summary: CRUD ACTION get_Class
      tags:
        - Subjects
      parameters:
        - name: class
          in: path
          description: Class
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
...
  /teacher/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Teachers
      parameters:
        - name: firstname
          in: path
          description: First Name
          required: true
          schema:
            type: string
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
        - name: lastname
          in: path
          description: Last Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Teachers
      parameters:
        - name: firstname
          in: path
          description: First Name
          required: true
          schema:
            type: string
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
        - name: lastname
          in: path
          description: Last Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teachers"
    post:
      summary: CRUD ACTION update
      tags:
        - Teachers
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teachers"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Teachers"
  /teacher/findBy_Class/{key}:
    get:
      summary: CRUD ACTION findBy_Class
      tags:
        - Teachers
      parameters:
        - name: class
          in: path
          description: Class
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teachers"
  /teacher/findBy_Exams/{key}:
    get:
      summary: CRUD ACTION findBy_Exams
      tags:
        - Teachers
      parameters:
        - name: examname
          in: path
          description: Exam Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teachers"
...
  /teacher/{id}/get_Class:
    get:
      summary: CRUD ACTION get_Class
      tags:
        - Teachers
      parameters:
        - name: class
          in: path
          description: Class
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
  /teacher/{id}/get_Exams:
    get:
      summary: CRUD ACTION get_Exams
      tags:
        - Teachers
      parameters:
        - name: examname
          in: path
          description: Exam Name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    
  #-----  Test  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Subjects:
      properties:
        _id:
          type: string
        averageresultinlastexam:
          type: number
        class:
          type: number
        name:
          type: string
        period:
          type: number
        teacher:
          type: string
      required:
        -  _id
        - class
        - name
        - teacher
    Teachers:
      properties:
        _id:
          type: string
        class:
          type: number
        firstname:
          type: string
        lastname:
          type: string
        middlename:
          type: string
      required:
        -  _id
        - class
        - firstname
        - lastname
...